Expect OFFICIAL_BUILD to be "true" for official builds
Both config/common.mk and config/version.mk expect OFFICIAL_BUILD
to be "true", not mearly set. We should have the same expectation
for build/envsetup.sh when setting the BUILD_NUMBER or warning about
it not being an official build.

(I made the mistake of using `export OFFICIAL_BUILD=1`.)

Change-Id: I5565be4fe1a65c1c4be43b10ff046415108fa742